requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.1
graphviz==0.20.3
psycopg2-binary==2.9.9
sqlparse==0.5.0
tzdata==2024.1
